refactor(iroh): store downloader as part of the blobs
